requirements.txt
fastapi==0.63.0
uvicorn==0.13.4
motor==2.3.1
python-decouple==3.4

